Decompose big component and slim stuff down
